requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==37.0.2
defusedxml==0.7.1
demjson3==3.0.5
Django==4.0.7
django-allauth==0.50.0
django-colorfield==0.5.0
django-compat==1.0.15
django-csp==3.7
django-feather==0.3.0
django-hosts==5.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-progressive-web-app==0.1.1
django4-background-tasks==1.2.7
future==0.18.2
idna==3.3
jsonfield==3.1.0
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==2.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2020.4
requests==2.26.0
requests-oauthlib==1.3.1
six==1.10.0
soupsieve==2.1
sqlparse==0.4.2
swapper==1.3.0
treelib==1.6.1
urllib3==1.26.7
